Paso de funcionalidades a signals.py
